created common selectors and updated few unit tests
