+++ src/components/FlowPanel/index.test.tsx
@@ -18,7 +18,7 @@ function setupContainer(
 ) {
   const customizedStore = {
     getState: () => ({
-      common: { showFLowPanel: false },
+      common: {},
       pairForm,
       terminals,
     }),
