+++ src/components/FlowPanel/index.tsx
@@ -3,6 +3,7 @@ import Box from '@material-ui/core/Box';
 import Drawer from '@material-ui/core/Drawer';
 import Typography from '@material-ui/core/Typography';
 import { useAppSelector } from '../../redux/hooks';
+import selectedShowFlowPanel from '../../redux/reducers/CommonSlice/commonSliceSelectors';
 import { selectPairFormSerialNumber } from '../../redux/reducers/PairFormSlice/PairFormSelectors';
 import { IPairingFlow, ITerminal, ITerminalProps } from '../../redux/reducers/TerminalSlice/interfaces';
 import { terminalInstance } from '../../redux/reducers/TerminalSlice/terminalsSliceSelectors';
@@ -12,7 +13,7 @@ export default function FlowPanel({ terminal }: ITerminal): React.ReactElement {
   const classes = useStyles();
   const pairFormSerialNumber = useAppSelector(selectPairFormSerialNumber);
   const currentTerminal = useAppSelector(terminalInstance(pairFormSerialNumber)) as ITerminalProps;
-  const { showFlowPanel } = useAppSelector((state) => state.common);
+  const showFlowPanel = useAppSelector(selectedShowFlowPanel);
 
   const statusInformation = (spi: ITerminalProps) => `
 # ----------- STATUS -----------
