+++ src/components/PairPage/index.tsx
@@ -2,6 +2,7 @@ import React, { useEffect } from 'react';
 import Container from '@material-ui/core/Container';
 import Grid from '@material-ui/core/Grid';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
+import selectedShowFlowPanel from '../../redux/reducers/CommonSlice/commonSliceSelectors';
 import { resetPairForm } from '../../redux/reducers/PairFormSlice/pairFormSlice';
 import FlowPanel from '../FlowPanel';
 import { IFlowPanel } from '../FlowPanel/interfaces';
@@ -13,7 +14,7 @@ import PairStatus from './PairStatus';
 
 const PairPage: React.FC = () => {
   const dispatch = useAppDispatch();
-  const { showFlowPanel } = useAppSelector((state) => state.common);
+  const showFlowPanel = useAppSelector(selectedShowFlowPanel);
   const classes = useStyles(showFlowPanel as unknown as IFlowPanel);
 
   useEffect(() => {
