+++ src/components/TerminalsPage/index.test.tsx
@@ -13,7 +13,7 @@ function setupContainer(
 ) {
   const customizedStore = {
     getState: () => ({
-      common: { showFLowPanel: false },
+      common: {},
       pairForm,
       terminals,
     }),
