+++ src/redux/reducers/CommonSlice/commonSliceSelectors.ts
@@ -1,0 +1,5 @@
+import { RootState } from '../../store';
+
+const selectedShowFlowPanel = (state: RootState): boolean => state.common.showFlowPanel;
+
+export default selectedShowFlowPanel;
