+++ src/redux/reducers/CommonSlice/interfaces/index.ts
@@ -1,0 +1,3 @@
+export interface ICommonState {
+  showFlowPanel: boolean;
+}
